re-tooling and refactoring
re-tooled the code to read/write the cartographer bin files, to be more usable in the creation of any bin file for any valid saved game fileset.

refactored out the code for the cartographer_profile into its own file and set of functions for usage.

rewrote the player control fov hook to use the new per-profile settings.
