+++ xlive/Project_Cartographer.vcxproj.filters
@@ -405,6 +405,7 @@
     <ClCompile Include="Blam\Engine\saved_games\saved_game_files.cpp" />
     <ClCompile Include="Blam\Engine\cache\phyical_memory_map.cpp" />
     <ClCompile Include="Blam\Engine\interface\user_interface_player_profile_list.cpp" />
+    <ClCompile Include="Blam\Engine\saved_games\cartographer_player_profile.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="3rdparty\miniupnpc\include\miniupnpc\codelength.h" />
@@ -1075,6 +1076,7 @@
     <ClInclude Include="Blam\Engine\cseries\async_helpers.h" />
     <ClInclude Include="Blam\Engine\cache\physical_memory_map.h" />
     <ClInclude Include="Blam\Engine\interface\user_interface_player_profile_list.h" />
+    <ClInclude Include="Blam\Engine\saved_games\cartographer_player_profile.h" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="resources\Resource.rc" />
